More robust error handling
Pingdom includes enough information in their error responses that we can
use to make very useful Cog error messages. Part of this includes
whether or not a given check identifier is valid, so we don't even need
to mess with the identifier value passed to the command as an argument.

(We can also remove that conversion from a string to a float to an
integer now that Relay has updated its JSON handling.)